+++ src/config/index.js
@@ -15,6 +15,6 @@ module.exports = {
             ADDRESS: 0x6f
         }
     },
-    SATOSHI: 100000000
-
+    SATOSHI: 100000000,
+    DEFAULT_SATOSHI_PER_BYTE: 20
 }